+++ coders/xwd.c
@@ -430,7 +430,7 @@ static Image *ReadXWDImage(const ImageInfo *image_info,ExceptionInfo *exception)
   /*
     Allocate the pixel buffer.
   */
-  length=(size_t) ximage->bytes_per_line*ximage->height;
+  length=(size_t) (ximage->bytes_per_line*ximage->height);
   if (CheckOverflowException(length,ximage->bytes_per_line,ximage->height))
     {
       if (header.ncolors != 0)
@@ -444,7 +444,7 @@ static Image *ReadXWDImage(const ImageInfo *image_info,ExceptionInfo *exception)
         extent;
 
       extent=length;
-      length*=ximage->depth;
+      length*=(size_t) ximage->depth;
       if (CheckOverflowException(length,extent,ximage->depth))
         {
           if (header.ncolors != 0)
@@ -544,16 +544,16 @@ static Image *ReadXWDImage(const ImageInfo *image_info,ExceptionInfo *exception)
             for (x=0; x < (ssize_t) image->columns; x++)
             {
               pixel=XGetPixel(ximage,(int) x,(int) y);
-              index=(Quantum) ConstrainColormapIndex(image,(ssize_t) (pixel >>
-                red_shift) & red_mask,exception);
+              index=(Quantum) ConstrainColormapIndex(image,(ssize_t) ((pixel >>
+                red_shift) & red_mask),exception);
               SetPixelRed(image,ScaleShortToQuantum(
                 colors[(ssize_t) index].red),q);
-              index=(Quantum) ConstrainColormapIndex(image,(ssize_t) (pixel >>
-                green_shift) & green_mask,exception);
+              index=(Quantum) ConstrainColormapIndex(image,(ssize_t) ((pixel >>
+                green_shift) & green_mask),exception);
               SetPixelGreen(image,ScaleShortToQuantum(
                 colors[(ssize_t) index].green),q);
-              index=(Quantum) ConstrainColormapIndex(image,(ssize_t) (pixel >>
-                blue_shift) & blue_mask,exception);
+              index=(Quantum) ConstrainColormapIndex(image,(ssize_t) ((pixel >>
+                blue_shift) & blue_mask),exception);
               SetPixelBlue(image,ScaleShortToQuantum(
                 colors[(ssize_t) index].blue),q);
               q+=GetPixelChannels(image);
